pyproject.toml
[project]
name = "first-adk"
version = "0.1.0"
description = ""
authors = [
    {name = "ninokin25",email = "masa25hw@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk (>=1.0.0,<2.0.0)",
    "litellm (>=1.71.1,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "first_adk", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
